Move RequireEqualResults and export for other tests to use
